requirements.txt
absl-py==0.6.1
appnope==0.1.0
astor==0.7.1
backcall==0.1.0
cycler==0.10.0
decorator==4.3.0
gast==0.2.0
grpcio==1.17.0
h5py==2.8.0
imageio==2.4.1
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.1
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.2
numpy==1.15.4
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prompt-toolkit==2.0.7
protobuf==3.6.1
ptyprocess==0.6.0
Pygments==2.3.0
pyparsing==2.3.0
python-dateutil==2.7.5
six==1.11.0
tensorboard==1.12.0
tensorflow==1.12.0
termcolor==1.1.0
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.14.1
